backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
fairlearn
scikit-learn
plotly
streamlit
matplotlib
seaborn
requests
huggingface_hub
python-dotenv
shap